+++ Tablet/Tablet.swift
@@ -106,11 +106,6 @@ public class Action {
     }
 }
 
-public class ActionObject<I, C> {
-    
-    
-}
-
 /**
     If you want to delegate your cell configuration logic to cell itself (with your view model or even model) than
     just provide an implementation of this protocol for your cell. Enjoy strong typisation.
@@ -135,163 +130,6 @@ public protocol RowBuilder {
     func triggerAction(key: String, cell: UITableViewCell, indexPath: NSIndexPath, itemIndex: Int) -> ActionResult
 }
 
-/**
-    A class that responsible for building cells of given type and passing items to them.
-*/
-public class TableRowBuilder<I, C where C: UITableViewCell> : RowBuilder {
-
-    public typealias ReturnValue = AnyObject
-
-    public typealias TableRowBuilderActionBlock = (data: ActionData<I, C>) -> Void
-    public typealias TableRowBuilderReturnValueActionBlock = (data: ActionData<I, C>) -> ReturnValue
-    
-    private var actions = Dictionary<String, TableRowBuilderActionBlock>()
-    private var items = [I]()
-
-    public var reusableIdentifier: String
-    public var numberOfRows: Int {
-        get {
-            return items.count
-        }
-    }
-
-    public init(item: I, id: String) {
-
-        reusableIdentifier = id
-        items.append(item)
-    }
-    
-    public init(items: [I]? = nil, id: String) {
-
-        reusableIdentifier = id
-        
-        if items != nil {
-            self.items.appendContentsOf(items!)
-        }
-    }
-
-    public func appendItems(items: [I]) {
-
-        self.items.appendContentsOf(items)
-    }
-    
-    public func clear() {
-
-        items.removeAll()
-    }
-
-    // MARK: Chaining actions
-
-    public func action(key: String, action: TableRowBuilderActionBlock) -> Self {
-
-        actions[key] = action
-        return self
-    }
-    
-    public func action(key: ActionType, action: TableRowBuilderActionBlock) -> Self {
-
-        actions[key.rawValue] = action
-        return self
-    }
-    
-    public func action(key: ActionType, action: TableRowBuilderReturnValueActionBlock) -> Self {
-        
-        
-        return self
-    }
-    
-    // MARK: Triggers
-
-    public func triggerAction(key: String, cell: UITableViewCell, indexPath: NSIndexPath, itemIndex: Int) -> ActionResult {
-
-        let actionData = ActionData(cell: cell as! C, indexPath: indexPath, item: items[itemIndex], itemIndex: itemIndex)
-
-        if let block = actions[key] {
-            block(data: actionData)
-            return .Failure
-        }
-        return .Failure
-    }
-}
-
-/**
-    A class that responsible for building configurable cells of given type and passing items to them.
-*/
-public class TableConfigurableRowBuilder<I, C: ConfigurableCell where C.Item == I, C: UITableViewCell> : TableRowBuilder<I, C>  {
-
-    public init(item: I) {
-        super.init(item: item, id: C.reusableIdentifier())
-    }
-
-    public init(items: [I]? = nil) {
-        super.init(items: items, id: C.reusableIdentifier())
-    }
-    
-    public override func triggerAction(key: String, cell: UITableViewCell, indexPath: NSIndexPath, itemIndex: Int) -> ActionResult {
-
-        (cell as! C).configureWithItem(items[itemIndex])
-        
-        return super.triggerAction(key, cell: cell, indexPath: indexPath, itemIndex: itemIndex)
-    }
-}
-
-/**
-    A class that responsible for building a certain table view section.
-*/
-public class TableSectionBuilder {
-
-    private var builders = [RowBuilder]()
-
-    public var headerTitle: String?
-    public var footerTitle: String?
-
-    public var headerView: UIView?
-    public var headerHeight: CGFloat = UITableViewAutomaticDimension
-    
-    public var footerView: UIView?
-    public var footerHeight: CGFloat = UITableViewAutomaticDimension
-
-    /// A total number of rows in section of each row builder.
-    public var numberOfRowsInSection: Int {
-
-        var number = 0
-        for builder in builders {
-            number += builder.numberOfRows
-        }
-        return number
-    }
-
-    public init(headerTitle: String? = nil, footerTitle: String? = nil, rowBuilders: [RowBuilder]? = nil) {
-
-        self.headerTitle = headerTitle
-        self.footerTitle = footerTitle
-        
-        if let initialRows = rowBuilders {
-            self.builders.appendContentsOf(initialRows)
-        }
-    }
-
-    public init(headerView: UIView? = nil, headerHeight: CGFloat = UITableViewAutomaticDimension, footerView: UIView? = nil, footerHeight: CGFloat = UITableViewAutomaticDimension) {
-
-        self.headerView = headerView
-        self.headerHeight = headerHeight
-
-        self.footerView = footerView
-        self.footerHeight = footerHeight
-    }
-
-    internal func builderAtIndex(var index: Int) -> (RowBuilder, Int)? {
-
-        for builder in builders {
-            if index < builder.numberOfRows {
-                return (builder, index)
-            }
-            index -= builder.numberOfRows
-        }
-
-        return nil
-    }
-}
 
 /**
     Responsible for table view's datasource and delegate.
